feat: rename,test add docs
