+++ deploy/examples/vllm/utils/vllm.py
@@ -45,6 +45,12 @@ def parse_vllm_args(service_name, prefix) -> AsyncEngineArgs:
         default=1000,
         help="Maximum length of local prefill",
     )
+    parser.add_argument(
+        "--cuda-visible-device-offset",
+        type=int,
+        default=0,
+        help="Offset of CUDA_VISIBLE_DEVICE",
+    )
     parser = AsyncEngineArgs.add_cli_args(parser)
     args = parser.parse_args(vllm_args)
     engine_args = AsyncEngineArgs.from_cli_args(args)
@@ -52,4 +58,5 @@ def parse_vllm_args(service_name, prefix) -> AsyncEngineArgs:
     engine_args.remote_prefill = args.remote_prefill
     engine_args.conditional_disagg = args.conditional_disagg
     engine_args.max_local_prefill_length = args.max_local_prefill_length
+    engine_args.cuda_visible_device_offset = args.cuda_visible_device_offset
     return engine_args
